requirements.txt
python-can==3.3.1
netifaces==0.10.4
pkcs7==0.1.2
sphinx-rtd-theme==0.4.2
#unittest
#termcolor
